celery/worker/requirements.txt
scrapy
fastapi
uvicorn
pandas
celery
redis>=4.4.0
pydantic
numpy
elasticsearch
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-propagator-gcp==1.3.0
opentelemetry-exporter-gcp-trace==1.3.0
opentelemetry-instrumentation-pika==0.41b0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-celery==0.41b0
opentelemetry-exporter-otlp==1.20.0
loguru
stackprinter